pyproject.toml
[project]
name = "local-voice-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # "fastrtc[stt]>=0.0.19",  # Using local editable version instead
    "gradio>=4.0.0",
    "kokoro-onnx>=0.4.7",
    "langchain-chroma>=0.1.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-ollama>=0.1.0",
    "langchain-text-splitters>=0.3.0",
    "librosa>=0.11.0",
    "loguru>=0.7.3",
    "numpy>=1.24.0",
    "ollama>=0.6.0",
    "openai-whisper>=20250625",
    "pypdf>=5.0.0",
    "qrcode[pil]>=7.4.2",
]
